refactor: move time related functions to a separate file
